Use case for image generation
